chore(docs): Xml docs for storage
