+++ src/Uno.Extensions.Storage/KeyValueStorage/KeyValueStorageExtensions.cs
@@ -1,13 +1,30 @@
 ﻿namespace Uno.Extensions.Storage.KeyValueStorage;
 
+/// <summary>
+/// Extensions for working with <see cref="IKeyValueStorage"/>.
+/// </summary>
 public static class KeyValueStorageExtensions
 {
+	/// <summary>
+	/// Determines if the storage contains the specified key.
+	/// </summary>
+	/// <param name="storage">The storage instance</param>
+	/// <param name="key">The key to search for</param>
+	/// <param name="ct">CancellationToken to cancel operation</param>
+	/// <returns>True if storage contains key</returns>
 	public static async ValueTask<bool> ContainsKey(this IKeyValueStorage storage, string key, CancellationToken ct)
 	{
 		var keys = await storage.GetKeysAsync(ct);
 		return keys.Contains(key);
 	}
 
+	/// <summary>
+	/// Gets a value from the storage as a string
+	/// </summary>
+	/// <param name="storage">The storage instance</param>
+	/// <param name="key">The key to retrieve</param>
+	/// <param name="ct">CancellationToken to cancel operation</param>
+	/// <returns>The value, or null if key not in storage</returns>
 	public static async ValueTask<string?> GetStringAsync(
 		this IKeyValueStorage storage,
 		string key,
@@ -21,16 +38,28 @@ public static class KeyValueStorageExtensions
 		return default;
 	}
 
+	/// <summary>
+	/// Retrieves all key/value pairs from the storage.
+	/// </summary>
+	/// <param name="storage">The storage instance</param>
+	/// <param name="ct">CancellationToken to cancel operation</param>
+	/// <returns>Dictionary of key-value pairs</returns>
 	public static ValueTask<IDictionary<string, string>> GetAllValuesAsync(
 		this IKeyValueStorage storage,
 		CancellationToken ct)
-	{
-		return GetAllValuesAsync(storage, _ => true, ct);
-	}
+			=> GetAllValuesAsync(storage, _ => true, ct);
+
+	/// <summary>
+	/// Retrieves all key/value pairs from the storage that match the predicate.
+	/// </summary>
+	/// <param name="storage">The storage instance</param>
+	/// <param name="predicate">The predicate to invoke to determine if pair should be returned</param>
+	/// <param name="ct">CancellationToken to cancel operation</param>
+	/// <returns>Dictionary of key-value pairs</returns>
 	public static async ValueTask<IDictionary<string, string>> GetAllValuesAsync(
-	this IKeyValueStorage storage,
-	Func<string, bool> predicate,
-	CancellationToken ct)
+		this IKeyValueStorage storage,
+		Func<string, bool> predicate,
+		CancellationToken ct)
 	{
 		var dict = new Dictionary<string, string>();
 		var keys = await storage.GetKeysAsync(ct);
@@ -45,13 +74,24 @@ public static class KeyValueStorageExtensions
 		return dict;
 	}
 
+	/// <summary>
+	/// Clear all values from the storage.
+	/// </summary>
+	/// <param name="storage">The storage instance</param>
+	/// <param name="ct">CancellationToken to cancel operation</param>
+	/// <returns>Task to await</returns>
 	public static ValueTask ClearAllAsync(
 	this IKeyValueStorage storage,
 	CancellationToken ct)
-	{
-		return ClearAllAsync(storage, _ => true, ct);	
-	}
+		=> ClearAllAsync(storage, _ => true, ct);
 
+	/// <summary>
+	/// Clear all values from storage that match the predicate.
+	/// </summary>
+	/// <param name="storage">The storage instance</param>
+	/// <param name="predicate">The predicate to invoke to determine if pair should be cleared</param>
+	/// <param name="ct">CancellationToken to cancel operation</param>
+	/// <returns>Task to await</returns>
 	public static async ValueTask ClearAllAsync(
 		this IKeyValueStorage storage,
 		Func<string, bool> predicate,
